Moved tests to tests
